[feat] Refactor project structure to accomadate for CLI
