+++ src/domain/error.rs
@@ -1,0 +1,7 @@
+pub type LoadBalancerResult<T> = Result<T, LoadBalancerError>;
+
+#[derive(Debug)]
+pub enum LoadBalancerError {
+    GenericError(Box<dyn std::error::Error>),
+    UnknowError,
+}
