+++ src/domain/mod.rs
@@ -1,0 +1,2 @@
+pub mod error;
+pub mod load_balancer;
