+++ src/lib.rs
@@ -1,0 +1,9 @@
+use domain::error::LoadBalancerResult;
+
+pub mod domain;
+
+mod cli;
+
+pub async fn run() -> LoadBalancerResult<()> {
+    cli::run().await
+}
